meson: generate and install headers and pkgconfig
This is able to do everything the autotools build does except generate
docbook. That shouldn't be too hard, I just didn't get to it.

meson requires a different format for the .h.in file, it doesn't
understand #undef, it uses #mesondef instead. The two files are exactly
the same except for that.